Feature: Role Authority 관계는 변경내용 기록할 필요 없으므로 BaseEntity상속하지 않고  ManyToMany로 구성

+++ lms-core/src/main/java/com/savelms/Authority.java
@@ -21,7 +21,7 @@ public class Authority {
 
     @Id
     @GeneratedValue
-    private Integer authorityId;
+    private Long authorityId;
 
     private String permission;
